Refactor GUIContentImp
