+++ src/main/java/com/github/webicitybrowser/thready/gui/graphical/base/imp/message/FocusManagerImp.java
@@ -1,4 +1,4 @@
-package com.github.webicitybrowser.thready.gui.graphical.base.imp;
+package com.github.webicitybrowser.thready.gui.graphical.base.imp.message;
 
 import java.util.Optional;
