+++ src/main/java/com/github/webicitybrowser/thready/windowing/core/ScreenContent.java
@@ -1,6 +1,8 @@
 package com.github.webicitybrowser.thready.windowing.core;
 
+import com.github.webicitybrowser.thready.dimensions.AbsolutePosition;
 import com.github.webicitybrowser.thready.dimensions.AbsoluteSize;
+import com.github.webicitybrowser.thready.dimensions.Rectangle;
 import com.github.webicitybrowser.thready.drawing.core.Canvas2D;
 import com.github.webicitybrowser.thready.drawing.core.ResourceLoader;
 import com.github.webicitybrowser.thready.gui.graphical.animation.InvalidationScheduler;
@@ -16,6 +18,10 @@ public interface ScreenContent {
 	
 	public static record ScreenContentRedrawContext(
 		Canvas2D canvas, AbsoluteSize contentSize, ResourceLoader resourceLoader, InvalidationScheduler invalidationScheduler
-	) {}
+	) {
+		public Rectangle rootDocumentRect() {
+			return new Rectangle(new AbsolutePosition(0, 0), contentSize);
+		}
+	}
 	
 }
